Introduced generalized Tosi-Fumi parameters and particle definitions
- Modified inputs
  - Added required input `N_typ`: number of atom types
  - Added required input `particles_file`: path to file defining atom species properties
  - Added optional input `potential_params_file`: path to file defining inter-atomic potential parameters (can be required for some potentials)
- Added optional input `tf_file`: path to file defining Tosi-Fumi parameters
- Changed `charges` to `double` instead of `long int`
- Can now initialize inter-atomic potentials with a set of parameters
- Files to define atom species with `TYPE, MASS, CHARGE, RADIUS`
- File to specify Tosi-Fumi parameters
  - Need one parameter for every possible pair of atoms types
  - Removed hard-coded TF parameters

+++ maze_poisson/c_api/get_functions.py
@@ -122,10 +122,10 @@ capi.register_function(
     ],
 )
 
-# void get_charges(long int *recv) {
+# void get_charges(double *recv) {
 capi.register_function(
     'get_charges', None, [
-        npct.ndpointer(dtype=np.int64, ndim=1, flags='C_CONTIGUOUS'),
+        npct.ndpointer(dtype=np.float64, ndim=1, flags='C_CONTIGUOUS'),
     ],
 )
